PageStatus moves do MoviesBrowser, imports updated
