--- src/common/Container/styled.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-const Container = styled.div`
-  max-width: 1920px;
-  margin: 0 auto;
-`;
-
-export default Container;
