+++ src/core/App/App.js
@@ -1,4 +1,4 @@
-import { PageStatus } from "../../features";
+import { PageStatus } from "../../features/MoviesBrowser/PageStatus";
 import Navigation from "../../common/Navigation";
 import { HashRouter, Switch, Route, Redirect } from "react-router-dom";
 import { toPeople } from "./routes";
@@ -6,12 +6,11 @@ import MovieDetails from "../../features/MoviesBrowser/MovieDetails";
 import PopularPeopleView from "../../features/MoviesBrowser/PopularPeopleView";
 import PeopleView from "../../features/MoviesBrowser/PeopleView";
 
-
 export const App = () => (
   <HashRouter>
     <Navigation />
     <Switch>
-    <Route path="/people/:id">
+      <Route path="/people/:id">
         <PeopleView />
       </Route>
       <Route path="/movies/:id">
